Tile Buttons Work
They work, but to clear past options. that method(clear options) is far
to processes intensive, should find better way, it currently goes three
deep. (only two off those can actually fluxucate). One fix could be to
keep a value of each option in memory in the container. But I think
there should be a way to do it dynamically. The trick is to figure out
how remove() works